Added seperate benchmark for large transform
